refactor: code refactoring

+++ backend/matechai/src/main/java/com/ft/matechai/user/node/User.java
@@ -55,7 +55,7 @@ public class User {
     @Builder.Default
     private List<String> pictureUrls = new ArrayList<>();
 
-    private int fame;
+    private Float fame;
     private String location;
 
     private String lastOnline;
